frontend/requirements.txt
Flask
python.dotenv
requests
request
flask-login
PyJWT
json